requirements.txt
fastapi
sqlmodel
uvicorn
pydantic[email]
alembic
pyJWT
bcrypt
passlib